moved configuration to the models folder, updated all the impacted files, updated the configuration as per the ticket
