--- src/configuration/index.ts
@@ -1,3 +1,0 @@
-// created from 'create-ts-index'
-
-export * from './configuration'
