+++ src/models/index.ts
@@ -11,6 +11,7 @@ export * from './block-meta'
 export * from './block'
 export * from './commit-signature'
 export * from './commit'
+export * from './configuration'
 export * from './consensus'
 export * from './node-params'
 export * from './node'
