+++ src/models/input/index.ts
@@ -1,7 +1,7 @@
 // created from 'create-ts-index'
 
-export * from './dispatch-request'
-export * from './node-proof'
-export * from './relay-payload'
-export * from './relay-request'
-export * from './session-header'
+export * from './dispatch-request';
+export * from './node-proof';
+export * from './relay-payload';
+export * from './relay-request';
+export * from './session-header';
