+++ src/utils/session-manager.ts
@@ -3,7 +3,7 @@ import { Session } from "../models/output/session"
 import { IKVStore } from "../utils/storage/kv-store"
 import { InMemoryKVStore } from "./storage/in-memory-kv-store"
 import { Node } from "../models"
-import { Configuration } from "../configuration/configuration"
+import { Configuration } from "../models/configuration"
 import { DispatchResponse } from "../models/output/dispatch-response"
 import { RpcErrorResponse } from "../models/output/rpc-error-response"
 import { RequestManager } from "../request-manager"
